pyproject.toml
[tool.poetry]
name = "dj-hotel"
version = "0.1.0"
description = ""
authors = ["surajmt8848 <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
black = "^21.5b1"
flake8 = "^3.9.2"
Pillow = "^8.2.0"
django-seed = "^0.2.2"
isort = "^5.8.0"
djangorestframework = "^3.12.4"
django-mptt = "^0.12.0"
django-extensions = "^3.1.3"
django-autoslug = "^1.9.8"
PyJWT = "^2.1.0"
drf-writable-nested = "^0.6.2"
drf-extensions = "^0.7.0"
django-hosts = "^4.0"
django-leaflet = "^0.28.0"
geopy = "^2.1.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
